+++ src/Encoding/ChunkStream.php
@@ -12,7 +12,7 @@ class ChunkStream extends FilteredStream
     /**
      * {@inheritdoc}
      */
-    public function getReadFilter()
+    protected function readFilter()
     {
         return 'chunk';
     }
@@ -20,7 +20,7 @@ class ChunkStream extends FilteredStream
     /**
      * {@inheritdoc}
      */
-    public function getWriteFilter()
+    protected function writeFilter()
     {
         return 'dechunk';
     }